+++ backend/src/main/java/com/woowacourse/zzimkkong/exception/map/MapException.java
@@ -1,4 +1,4 @@
-package com.woowacourse.zzimkkong.exception;
+package com.woowacourse.zzimkkong.exception.map;
 
 import org.springframework.http.HttpStatus;
