+++ backend/src/main/java/com/woowacourse/zzimkkong/exception/reservation/ReservationException.java
@@ -1,11 +1,11 @@
-package com.woowacourse.zzimkkong.exception;
+package com.woowacourse.zzimkkong.exception.reservation;
 
 import org.springframework.http.HttpStatus;
 
-public class SpaceException extends RuntimeException {
+public class ReservationException extends RuntimeException {
     private final HttpStatus status;
 
-    public SpaceException(final String message, final HttpStatus status) {
+    public ReservationException(final String message, final HttpStatus status) {
         super(message);
         this.status = status;
     }
