+++ backend/src/main/java/com/woowacourse/zzimkkong/service/AuthService.java
@@ -3,8 +3,8 @@ package com.woowacourse.zzimkkong.service;
 import com.woowacourse.zzimkkong.domain.Member;
 import com.woowacourse.zzimkkong.dto.member.LoginRequest;
 import com.woowacourse.zzimkkong.dto.member.TokenResponse;
-import com.woowacourse.zzimkkong.exception.NoSuchMemberException;
-import com.woowacourse.zzimkkong.exception.PasswordMismatchException;
+import com.woowacourse.zzimkkong.exception.member.NoSuchMemberException;
+import com.woowacourse.zzimkkong.exception.member.PasswordMismatchException;
 import com.woowacourse.zzimkkong.infrastructure.JwtUtils;
 import com.woowacourse.zzimkkong.repository.MemberRepository;
 import org.springframework.stereotype.Service;
