+++ backend/src/test/java/com/woowacourse/zzimkkong/repository/MapRepositoryTest.java
@@ -2,7 +2,7 @@ package com.woowacourse.zzimkkong.repository;
 
 import com.woowacourse.zzimkkong.domain.Map;
 import com.woowacourse.zzimkkong.domain.Member;
-import com.woowacourse.zzimkkong.exception.NoSuchMapException;
+import com.woowacourse.zzimkkong.exception.map.NoSuchMapException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
